Renames Project -> ProjectListItem
