+++ src/pages/app/sidebar/projects/ProjectListItem.svelte
